Refactor: `colonyActions` into their own resolver
